pyproject.toml
[project]
name = "bicameral"
version = "2.0.0"
description = "AI collaboration infrastructure for Claude & Gemini"
authors = [{name = "Team RADIORHINO", email = "hello@rhncrs.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ai", "collaboration", "redis", "claude", "gemini"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/your-org/bicameral"
Documentation = "https://github.com/your-org/bicameral/tree/main/docs"
Repository = "https://github.com/your-org/bicameral"
Issues = "https://github.com/your-org/bicameral/issues"

[project.scripts]
bicameral = "bicameral.__main__:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bicameral"]
package-dir = {"" = "scripts"}

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
